@@ -10,9 +10,7 @@ __all__ = [
     'byte2int',
     'bytes2bits',
     'bytes2integer',
-    'bytes2integers',
     'bytes2str',
-    'bytestringtype',
     'Container',
     'globalPrintFalseFlags',
     'globalPrintFullStrings',
@@ -27,26 +25,29 @@ __all__ = [
     'int2byte',
     'integer2bits',
     'integer2bytes',
-    'integers2bytes',
-    'integertypes',
     'ListContainer',
     'ONWINDOWS',
     'PY',
-    'PY2',
-    'PY3',
     'PYPY',
     'RebufferedBytesIO',
-    'reprstring',
     'RestreamedBytesIO',
     'setGlobalPrintFalseFlags',
     'setGlobalPrintFullStrings',
     'setGlobalPrintPrivateEntries',
     'str2bytes',
-    'stringtypes',
     'swapbitsinbytes',
     'swapbytes',
     'swapbytesinbits',
-    'trimstring',
     'unhexlify',
+    # deprecated:
+    'PY2',
+    'PY3',
+    'bytestringtype',
+    'bytes2integers',
+    'integers2bytes',
+    'integertypes',
+    'reprstring',
+    'stringtypes',
     'unicodestringtype',
+    'trimstring',
 ]
